+++ .babelrc
@@ -1,21 +1,22 @@
 {
-  "presets": ["react", "env"],
-  "plugins": [
-    "transform-object-rest-spread",
-    "transform-class-properties",
-    "transform-runtime"
-  ],
-  "env": {
-    "development": {
-      "plugins": [
-        ["react-transform", {
-          "transforms": [{
-            "transform": "react-transform-hmr",
-            "imports": ["react"],
-            "locals": ["module"]
-          }]
-        }]
-      ]
+    "presets": ["react", "env"],
+    "plugins": ["transform-object-rest-spread", "transform-class-properties", "transform-runtime"],
+    "env": {
+        "development": {
+            "plugins": [
+                [
+                    "react-transform",
+                    {
+                        "transforms": [
+                            {
+                                "transform": "react-transform-hmr",
+                                "imports": ["react"],
+                                "locals": ["module"]
+                            }
+                        ]
+                    }
+                ]
+            ]
+        }
     }
-  }
 }
