+++ package.json
@@ -13,7 +13,6 @@
         "lint": "eslint src test playground",
         "prepare": "npm run dist",
         "precommit": "lint-staged",
-        "publish-to-gh-pages": "npm run build:playground && gh-pages --dist build/",
         "publish-to-npm": "npm run build:readme && npm run dist && npm publish",
         "preversion": "npm run build:playground && npm run dist && npm run build:readme && npm run cs-check && npm run lint",
         "start": "node devServer.js",
@@ -42,8 +41,7 @@
     },
     "dependencies": {},
     "devDependencies": {
-        "@ver01/form-theme-antd": "^1.0.1",
-        "antd": "^3.12.4",
+        "@ver01/form-theme-antd": "^1.0.2",
         "atob": "^2.0.3",
         "babel-cli": "^6.18.0",
         "babel-core": "^6.26.3",
@@ -59,7 +57,7 @@
         "chai": "^3.3.0",
         "codemirror": "^5.30.0",
         "cross-env": "^2.0.1",
-        "css-loader": "^0.23.1",
+        "css-loader": "^2.1.1",
         "eslint": "^4.9.0",
         "eslint-config-react-app": "^2.0.1",
         "eslint-plugin-flowtype": "^2.39.1",
@@ -69,7 +67,6 @@
         "estraverse": "^4.2.0",
         "estraverse-fb": "^1.3.1",
         "express": "^4.14.0",
-        "gh-pages": "^0.11.0",
         "html": "^1.0.0",
         "husky": "^0.13.2",
         "jsdom": "^8.3.0",
@@ -77,9 +74,7 @@
         "lint-staged": "^3.3.1",
         "mini-css-extract-plugin": "^0.4.3",
         "mocha": "^5.2.0",
-        "moment": "^2.24.0",
         "prettier": "^1.16.4",
-        "rc-color-picker": "^1.2.6",
         "react": "^15.5.0",
         "react-addons-test-utils": "^15.3.2",
         "react-codemirror2": "^4.1.0",
@@ -89,7 +84,7 @@
         "redbox-react": "^1.3.3",
         "rimraf": "^2.5.4",
         "sinon": "^1.17.6",
-        "style-loader": "^0.13.1",
+        "style-loader": "^0.23.1",
         "toctoc": "^0.2.3",
         "webpack": "^4.20.2",
         "webpack-cli": "^3.1.2",
