+++ playground/index.prod.html
@@ -1,18 +1,23 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>@ver01/form playground</title>
-  <link rel="stylesheet" href="./styles.css">
-</head>
-<body>
-  <div id="app"></div>
-  <p style="text-align:center">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>@ver01/form playground</title>
+    <link rel="stylesheet" href="./styles.css">
+    <link
+      rel="stylesheet"
+      id="theme"
+      href="https://cdn.bootcss.com/twitter-bootstrap/3.3.6/css/bootstrap.min.css"
+    />
+  </head>
+  <body>
+    <div id="app"></div>
+    <p style="text-align:center">
     Powered by
     <a href="https://github.com/ver01/form">@ver01/form</a>.
   </p>
-  <script src="./bundle.js"></script>
-</body>
+    <script src="./bundle.js"></script>
+  </body>
 </html>
