+++ playground/schemaDemo/arrays.js
