+++ playground/schemaDemo/customArray.js
